feat: validate methods

+++ student/api.py
@@ -1,11 +1,12 @@
 from typing import List
 
-from django.shortcuts import get_object_or_404
 from ninja import Router
+from ninja.errors import HttpError
+
 from calification.models import Calification
 from core.utils import get_letter_calification
-
 from teacher.bearer import AuthBearer
+from teacher.schemas.response import ResponseMessage
 
 from .constants import Endpoints
 from .models import Student
@@ -46,48 +47,58 @@ def get_student(request, id):
 @router.post(
     Endpoints.POST_ADD_STUDENT,
     auth=AuthBearer(),
-    response={201: str},
+    response={201: ResponseMessage},
 )
 def add_student(request, data: List[PayloadPostAddStudent], course_id):
     """
     Create a new politician.
     """
-    for payload in data:
-        student_data = payload.dict()
-        student_data['course_id'] = course_id
-        Student.objects.create(**student_data)
-
-    return "all students were added"
+    try:
+        for payload in data:
+            student_data = payload.dict()
+            student_data['course_id'] = course_id
+            Student.objects.create(**student_data)
+        return dict(message="All students were added")
+    except:
+        raise HttpError(403, "Cannot register student")
 
 
 @router.put(
     Endpoints.PUT_STUDENT,
     auth=AuthBearer(),
-    response={201: str},
+    response={201: ResponseMessage},
 )
 def put_my_student(request, data: PayloadUpdateStudent, id):
     """
     Edit my student data.
     """
-    student = get_object_or_404(Student, id=id)
-    student_data = data.dict(exclude_unset=True)
+    try:
+        student = Student.objects.get(id=id)
+        if student:
+            student_data = data.dict(exclude_unset=True)
 
-    for key, value in student_data.items():
-        setattr(student, key, value)
+            for key, value in student_data.items():
+                setattr(student, key, value)
 
-    student.save()
-    return f'Course {student.name} was updated'
+            student.save()
+            return dict(message=f'Student {student.first_name} was updated')
+    except:
+        raise HttpError(403, "Student not found")
 
 
 @router.delete(
     Endpoints.DELETE_STUDENT,
     auth=AuthBearer(),
-    response={204: str},
+    response={204: ResponseMessage},
 )
 def delete_student(request, id):
     """
     Delete student.
     """
-    student = get_object_or_404(Student, id=id)
-    student.delete()
-    return f'Student #{id} was deleted'
+    try:
+        student = Student.objects.get(id=id)
+        if student:
+            student.delete()
+            return dict(message=f'Student {student.first_name} was deleted')
+    except:
+        raise HttpError(403, "Student not found")
